Upgrade to flutter 3.10 and dart 3

+++ lib/app/modules/home/elements/news_carousel.dart
@@ -436,7 +436,7 @@ class _AppNewsCarouselState extends State<AppNewsCarousel> {
             fit: BoxFit.scaleDown,
             child: DotsIndicator(
               dotsCount: pages.length,
-              position: _curCarouselIndex.toDouble(),
+              position: _curCarouselIndex,
               decorator: const DotsDecorator(
                 color: Colors.white70,
                 spacing: EdgeInsets.symmetric(vertical: 6, horizontal: 3),
